refactor: Navigation.jsx & CommonLayout.jsx 이동
