+++ src/pages/menu/MenuDetail.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { useNavigate, useLocation } from 'react-router-dom';
-import CommonLayout from '../../layouts/CommonLayout';
+import CommonLayout from '../../components/layouts/CommonLayout';
 import MenuHeader from '../../components/menu/MenuHeader';
 import MenuInfo from '../../components/menu/MenuInfo';
 import OrderActionBtn from '../../components/menu/menu_detail/MenuDetailFooter';
